Change License MIT to MPL w/ Healthcare Disclaimer
OpenMRS repos use the MPL v2 with Healthcare Disclaimer. We believe this repo was unintentionally set up with MIT by mistake. I've consulted with Burke & Daniel, and we're proposing we update the License in this repo to use the exact same license as this one (text copy/pasted): https://github.com/openmrs/openmrs-core/blob/master/LICENSE